backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
black==24.10.0
click==8.1.7
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
isort==5.13.2
Markdown==3.7
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pydantic==2.10.4
pydantic_core==2.27.2
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
